Added gold cap and safe cd
Also a rob_def user command.

+++ shop/shop.py
@@ -465,7 +465,14 @@ class Shop:
             except RuntimeError:
                 return
 
+        
         user_data = await self.get_instance(ctx, user=ctx.author)
+        # Gold Bar Block #1. Another in sm.order()
+        if item == "Gold Bar":
+            inventory = await user_data.Inventory.all()
+            if item in inventory:
+                return await ctx.send("You can only have **1** Gold Bar!")
+
         sm = ShopManager(ctx, instance, user_data)
         try:
             await sm.order(shop, item)
@@ -1132,8 +1139,9 @@ class ShopManager:
         cur = await bank.get_currency_name(self.ctx.guild)
         stock, cost, _type = item_data['Qty'], item_data['Cost'], item_data['Type']
 
+        # Edited in "Gold Bar" statement. -Zyl
         await self.ctx.send("How many {} would you like to purchase?\n*If this "
-                            "is a random item, you can only buy 1 at a time.*".format(item))
+                            "is a random item or a Gold Bar, you can only buy 1 at a time.*".format(item))
 
         def predicate(m):
             if m.author == self.ctx.author and m.content.isdigit():
@@ -1147,6 +1155,12 @@ class ShopManager:
                 return False
         num = await self.ctx.bot.wait_for('message', timeout=25.0, check=predicate)
         amount = int(num.content)
+        print(amount)
+        print(item)
+        # Gold Bar Block #2
+        if item == "Gold Bar" and amount > 1:
+            return await self.ctx.send("You can only purchase **1** Gold Bar!")
+        #
         try:
             await num.delete()
         except discord.NotFound:
